more AI changes
the AI should be just about done.

I need to finish creating the prefabs for enemies when I am more awake.

Once I have created the prefab you can just drag the enemy in to the scene, give it some checkpoints to patrol and it should work.

Player and AI still need a tiny bit more work to make them fully functional for the game, and player anims still need to be updated